:recycle: Move all generators to separate files
